# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Roombnb API",
        "description": "API para listagem de rooms, criação de reservas e listagem de reservas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Ambiente local"
        }
    ],
    "paths": {
...
        "/api/v1/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Lista rooms disponíveis",
                "operationId": "31a8839ec4012ed828bcc67b19738698",
                "parameters": [
                    {
                        "name": "city_id",
                        "in": "query",
                        "description": "Filtrar por cidade",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "specialty_id",
                        "in": "query",
                        "description": "Filtrar por especialidade",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "check_in",
                        "in": "query",
                        "description": "Data de check-in (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-02-11"
                        }
                    },
                    {
                        "name": "check_out",
                        "in": "query",
                        "description": "Data de check-out (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-02-12"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de rooms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Consultório Centro"
                                                    },
                                                    "price": {
                                                        "type": "integer",
                                                        "example": 15000
                                                    },
                                                    "rating_avg": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "cover_url": {
                                                        "type": "string",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/rooms/{room}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Exibe os detalhes completos de um room",
                "operationId": "3aa744ef2227fff59fd43e569e482291",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "description": "ID do room",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do room",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Consultório no Centro"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Sala climatizada com recepção."
                                                },
                                                "city": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 3
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Salvador"
                                                        },
                                                        "state": {
                                                            "type": "string",
                                                            "example": "BA"
                                                        }
                                                    },
                                                    "type": "object",
                                                    "nullable": true
                                                },
                                                "specialty": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 7
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Fisioterapia"
                                                        }
                                                    },
                                                    "type": "object",
                                                    "nullable": true
                                                },
                                                "price": {
                                                    "description": "Preço em centavos",
                                                    "type": "integer",
                                                    "example": 15000
                                                },
                                                "rating_avg": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "cover_url": {
                                                    "type": "string",
                                                    "example": "https://site.com/storage/rooms/abc.jpg",
                                                    "nullable": true
                                                },
                                                "pictures": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 55
                                                            },
                                                            "url": {
                                                                "type": "string",
                                                                "example": "https://site.com/storage/rooms/pic1.jpg"
                                                            },
                                                            "is_cover": {
                                                                "type": "boolean",
                                                                "example": false
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Room not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token JWT ou Sanctum no formato: Bearer {token}"
            }
        }
    },
...
}